+++ database/migrations/2022_05_10_132249_create_email_template_descriptions_table.php
@@ -13,15 +13,18 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('email_template_descriptions', function (Blueprint $table) {
-            $table->id();
-            $table->integer('parent_id')->nullable();
-            $table->integer('language_id')->nullable();    
-           $table->string('name')->nullable();
-            $table->string('subject')->nullable();
-            $table->text('body')->nullable();
-            $table->timestamps();
-        });
+        // Check if the email_template_descriptions table exists
+        if (!Schema::hasTable('email_template_descriptions')) {
+            Schema::create('email_template_descriptions', function (Blueprint $table) {
+                $table->id();
+                $table->integer('parent_id')->nullable();
+                $table->integer('language_id')->nullable();
+                $table->string('name')->nullable();
+                $table->string('subject')->nullable();
+                $table->text('body')->nullable();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -31,6 +34,9 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('email_template_descriptions');
+        // Check if the email_template_descriptions table exists before attempting to drop it
+        if (Schema::hasTable('email_template_descriptions')) {
+            Schema::dropIfExists('email_template_descriptions');
+        }
     }
 };
