+++ database/migrations/2022_05_19_102831_create_designation_permission_actions_table.php
@@ -13,16 +13,19 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('designation_permission_actions', function (Blueprint $table) {
-            $table->id();
-            $table->integer('designation_id')->nullable();
-            $table->integer('designation_permission_id')->nullable();
-            $table->integer('admin_module_id')->nullable();    
-            $table->string('admin_sub_module_id')->nullable();
-            $table->integer('admin_module_action_id')->nullable();
-            $table->integer('is_active')->nullable();   
-            $table->timestamps();
-        });
+        // Check if the designation_permission_actions table exists
+        if (!Schema::hasTable('designation_permission_actions')) {
+            Schema::create('designation_permission_actions', function (Blueprint $table) {
+                $table->id();
+                $table->integer('designation_id')->nullable();
+                $table->integer('designation_permission_id')->nullable();
+                $table->integer('admin_module_id')->nullable();
+                $table->string('admin_sub_module_id')->nullable();
+                $table->integer('admin_module_action_id')->nullable();
+                $table->integer('is_active')->default(1); // Added default value
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -32,6 +35,9 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('designation_permission_actions');
+        // Check if the designation_permission_actions table exists before attempting to drop it
+        if (Schema::hasTable('designation_permission_actions')) {
+            Schema::dropIfExists('designation_permission_actions');
+        }
     }
 };
