+++ database/migrations/2022_05_27_063748_create_lookups_table.php
@@ -3,6 +3,7 @@
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Support\Facades\DB;
 
 return new class extends Migration
 {
@@ -13,13 +14,16 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('lookups', function (Blueprint $table) {
-            $table->id();
-            $table->string('code');
-            $table->string('lookup_type');
-            $table->unsignedBigInteger('is_active')->default('1');
-            $table->timestamps();
-        });
+        // Check if the table already exists
+        if (!Schema::hasTable('lookups')) {
+            Schema::create('lookups', function (Blueprint $table) {
+                $table->id();
+                $table->string('code');
+                $table->string('lookup_type');
+                $table->boolean('is_active')->default(true); // Changed to boolean
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -29,6 +33,9 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('lookups');
+        // Optionally, you might want to drop the table if it exists
+        if (Schema::hasTable('lookups')) {
+            Schema::dropIfExists('lookups');
+        }
     }
 };
